+++ Backend API/src/Repositories/departments.repository.js
@@ -1,18 +1,7 @@
-const { FatalError, BadRequestError, NotImplementedError } = require("../Core/Abstractions/Exceptions");
-const { isNullUndefinedOrEmpty, isNullOrUndefined, isListEmpty } = require("../Core/Utils/null-checker.util");
+const { BadRequestError, NotImplementedError } = require("../Core/Abstractions/Exceptions");
+const { isNullUndefinedOrEmpty, isNullOrUndefined } = require("../Core/Utils/null-checker.util");
 
-const DatabaseManager = require("../Database/database");
-const { Department } = require('./Entities/index');
-
-const tableName = "departments";
-
-const createDepartment = (description) => {
-    if(isNullUndefinedOrEmpty(description)){
-        throw new BadRequestError('Description cannot be empty');
-    }
-
-    return DatabaseManager.run(`INSERT INTO ${tableName} (description) VALUES ('${description}')`);
-};
+const Department = require('./Entities/department.class');
 
 const createDepartmentAsync = (description) => {
     if(isNullUndefinedOrEmpty(description)){
@@ -22,15 +11,6 @@ const createDepartmentAsync = (description) => {
     return Department.create({
         description
     });
-}
-
-const getDepartmentById = (id) => {
-    let departments = DatabaseManager.query(`SELECT * FROM ${tableName} WHERE id = ${+id} LIMIT 1`);
-    if (isListEmpty(departments)) {
-        return undefined;
-    }
-
-    return departments[0];
 };
 
 const getDepartmentByIdAsync = async (id) => {
@@ -40,12 +20,6 @@ const getDepartmentByIdAsync = async (id) => {
     }
 
     return department;
-}
-
-const getDepartments = () => {
-    let departments = DatabaseManager.query(`SELECT * FROM ${tableName}`);
-
-    return departments;
 };
 
 const getDepartmentsAsync = (skip = 0, limit = 10, orderBy = 'DESC') => Department.findAll({
@@ -53,28 +27,6 @@ const getDepartmentsAsync = (skip = 0, limit = 10, orderBy = 'DESC') => Departme
     offset: skip,
     limit
 });
-
-const updateDepartment = (id, newDescription) => {
-    if(isNullUndefinedOrEmpty(newDescription)){
-        throw new BadRequestError('Description cannot be empty');
-    }
-
-    let department = getDepartmentById(id);
-    if (isNullOrUndefined(department)) {
-        return undefined;
-    }
-
-    if (department.description === newDescription) {
-        return department;
-    }
-
-    let result = DatabaseManager.run(`UPDATE ${tableName} SET description = '${newDescription}' WHERE id = ${id}`);
-    if (result.changes === 0) {
-        throw new FatalError(`Unable to update department with id '${id}'`);
-    }
-
-    return getDepartmentById(id);
-};
 
 const updateDepartmentAsync = async (id, newDescription) => {
     if(isNullUndefinedOrEmpty(newDescription)){
@@ -102,13 +54,9 @@ const deleteDepartmentAsync = (id) => {
 }
 
 module.exports = {
-    createDepartment,
     createDepartmentAsync,
-    getDepartmentById,
     getDepartmentByIdAsync,
-    getDepartments,
     getDepartmentsAsync,
-    updateDepartment,
     updateDepartmentAsync,
     deleteDepartmentAsync,
 }
